+++ packages/pillar-ui/src/components/drawer/index.ts
@@ -1,0 +1,11 @@
+import type { App } from 'vue';
+import Drawer from './Drawer.vue';
+import './style.css';
+
+export { Drawer };
+
+export const XxDrawer = Object.assign(Drawer, {
+  install(app: App): void {
+    app.component('XxDrawer', Drawer);
+  },
+});